+++ migrations/2_deploy_contracts.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 const Hub = artifacts.require('./Hub.sol');
 const { convertToBaseUnit } = require('../test/helpers/math');
 
@@ -11,5 +13,10 @@ module.exports = async function (deployer) {
     convertToBaseUnit(50),
     '92592592592592',
     '7776000',
-  );
+  ).then((result) => {
+    fs.writeFile('addresses', `${result.address} \n`, (err) => {
+      // Checking for errors
+      if (err) throw err;
+    });
+  });
 };