+++ test/helpers/executeSafeTx.js
@@ -9,12 +9,13 @@ const executeSafeTx = async (safe, txParams, from, baseGas, safeTxGas, signer, w
   const operation = txParams.operation || 0;
   // gas parameters that were signed don't have to match those sent with the transaction
   // in cases where not using a token payment
-  const signedSafeTxGas = txParams.safeTxGas || 0;
+  const signedSafeTxGas = txParams.safeTxGas || 1;
   const signedBaseGas = txParams.baseGas || 0;
   const gasPrice = txParams.gasPrice || 0;
   const gasToken = txParams.gasToken || ZERO_ADDRESS;
   const refundReceiver = txParams.refundReceiver || ZERO_ADDRESS;
   const nonce = (await safe.nonce()).toNumber();
+  const chainId =   await web3.eth.getChainId();
 
   // this is mirroring the safe contracts gas checks inside execTransaction that account for EIP-150
   const max = Math.floor(Math.max((safeTxGas * 64) / 63, safeTxGas + 2500) + 500);
@@ -22,7 +23,7 @@ const executeSafeTx = async (safe, txParams, from, baseGas, safeTxGas, signer, w
 
   const typedData = formatTypedData(
     to, value, data, operation, signedSafeTxGas, signedBaseGas, gasPrice,
-    gasToken, refundReceiver, nonce, safe.address);
+    gasToken, refundReceiver, nonce, chainId, safe.address);
 
   const signatureBytes = await signTypedData(signer, typedData, web3);
   await safe.execTransaction(